Almost completely rewritten the code. The only untouched things are 2D convex hull and akl-tou heuristic.
The idea of the changes is more clever data representation and resource
allocation:
- In data representation I got rid of all the HashSets and Dictionaries in
favor of Lists and started using "mark&sweep" approach to enumerate unique
vertices.
- In the old version there was a lof of throwaway objects created. Now, the
algorithm recycles the faces that would otherwise be thrown away.
- Replaced all "important" foreach blocks with for loops. This alone made
the algorithm a lot faster due to reduced strain on the Garbage Collector.
- A lot of small tweaks.

As a result, the algorithm is now significantly faster. For 3D triangulation,
it is 2-5 times faster (6s for bunny model down from 23s of the previous
version) and much more memory efficient (the peak memory used by the algorithm
got reduced by about 5 times (not in the case of 2D hull)).

Also, the API got completely revamped. The usage should be obvious from the
examples.

This commit does not solve the issue when triangulating "regular" data (data
that contain a regular lattice). However, introducing small randomness to the
data usually solves the problem and the data gets triangulated just fine.
This (somewhat) is illustrated in Example 6 (Using Helix) where all the models
can now be triangulated.

Also, the Status class introduced by Matt in last commit isn't present here.